User settings creation upon sign up
For the quick addition of pairing users to classes the views file will check to see if a "userSettings" folder exist if not it will create one then create a txt file with the user's name.

Currently this does not pair the text file with the django database so if I user account is deleted the file will still linger and won't be directly edited though django.

Should also be changed if we keep this method to only create a userSetting's txt file if an account is verified.